Dumping
- Must specify pod id during registration
- Use node id instead of name
- bunch of other refactors too much stuff I don't even remember

+++ src/routers/node.py
@@ -2,8 +2,9 @@ from fastapi import APIRouter, Depends, BackgroundTasks
 import requests
 
 from src.internal.type import Resp, WsType
-from src.internal.manager import clusters, update
+from src.internal.manager import manager, update
 from src.internal.auth import verify_setup
+from src.utils.config import clusters
 
 
 router = APIRouter(tags=["node"])
@@ -12,6 +13,7 @@ router = APIRouter(tags=["node"])
 @router.get("/cloud/node/", dependencies=[Depends(verify_setup)])
 async def node_ls(pod_id: str | None = None) -> Resp:
     """monitoring: 2. cloud node ls [RES_POD_ID]"""
+    # FIXME: need to know which cluster has that pod
     return Resp.parse_raw(
         requests.get(
             clusters["5551"] + "/cloud/node/", params={"pod_id": pod_id}
@@ -21,39 +23,48 @@ async def node_ls(pod_id: str | None = None) -> Resp:
 
 @router.post("/cloud/node/", dependencies=[Depends(verify_setup)])
 async def node_register(
-    background_tasks: BackgroundTasks, node_name: str, pod_id: str | None = None
+    background_tasks: BackgroundTasks, node_name: str, pod_id: str
 ) -> Resp:
-    """management: 4. cloud register NODE_NAME [POD_ID]"""
+    """management: 4. cloud register NODE_NAME POD_ID"""
     if len(node_name) >= 16:
         return Resp(status=False, msg="manager: node name is too long!")
-    resp = Resp.parse_raw(
-        requests.post(
-            clusters["5551"] + "/cloud/node/",
-            params={"node_name": node_name, "pod_id": pod_id},
-        ).content
-    )
+    cluster = manager.pod_map.get(pod_id)
+    if cluster is None:
+        return Resp(status=False, msg=f"manager: pod_id {pod_id} not found")
+    resp = requests.post(
+        clusters[cluster] + "/cloud/node/",
+        params={"node_name": node_name, "pod_id": pod_id},
+    ).json()
+    if resp["status"] == False:
+        return Resp(status=False, msg=resp["msg"])
+    manager.node_map[resp["data"]] = cluster
     background_tasks.add_task(update, WsType.POD)
     background_tasks.add_task(update, WsType.NODE)
-    return resp
+    return Resp(status=True, data=resp["msg"], msg=resp["msg"])
 
 
 @router.delete("/cloud/node/", dependencies=[Depends(verify_setup)])
-async def node_rm(background_tasks: BackgroundTasks, node_name: str) -> Resp:
+async def node_rm(background_tasks: BackgroundTasks, node_id: str) -> Resp:
     """management: 5. cloud rm NODE_NAME"""
-    resp = Resp.parse_raw(
-        requests.delete(
-            clusters["5551"] + "/cloud/node/",
-            params={"node_name": node_name},
-        ).content
-    )
+    cluster = manager.node_map.get(node_id)
+    if cluster is None:
+        return Resp(status=False, msg=f"manager: node_id {node_id} not found")
+    resp = requests.delete(
+        clusters[cluster] + "/cloud/node/",
+        params={"node_id": node_id},
+    ).json()
+    if resp["status"] == False:
+        return Resp(status=False, msg=resp["msg"])
+    manager.node_map.pop(node_id)
     background_tasks.add_task(update, WsType.POD)
     background_tasks.add_task(update, WsType.NODE)
-    return resp
+    return Resp(status=True, msg=resp["msg"])
 
 
 @router.get("/cloud/node/log/", dependencies=[Depends(verify_setup)])
 async def node_log(node_id: str) -> Resp:
     """monitoring: 5. cloud node log NODE_ID"""
+    # FIXME: need to know which cluster to get the node log
     return Resp.parse_raw(
         requests.get(
             clusters["5551"] + "/cloud/node/log/",
